# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Headmaster API - Personal AI Knowledge System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InputType = (typeof InputType)[keyof typeof InputType];

export const InputType = {
  text: "text",
  url: "url",
  note: "note",
  file: "file",
  youtube: "youtube",
  instagram: "instagram",
  pinterest: "pinterest",
  image: "image",
  video: "video",
  audio: "audio",
} as const;

export interface Input {
  id: string;
  title: string;
  content: string;
  type: InputType;
  summary?: string | null;
  processed: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateInputBodyType =
  (typeof CreateInputBodyType)[keyof typeof CreateInputBodyType];

export const CreateInputBodyType = {
  text: "text",
  url: "url",
  note: "note",
  file: "file",
  youtube: "youtube",
  instagram: "instagram",
  pinterest: "pinterest",
  image: "image",
  video: "video",
  audio: "audio",
} as const;

export interface CreateInputBody {
  title: string;
  content: string;
  type: CreateInputBodyType;
}

export interface PlaylistEntry {
  title: string;
  videoId?: string | null;
  hasTranscript: boolean;
}

export interface YoutubeImportResult {
  input: Input;
  playlistEntries?: PlaylistEntry[] | null;
}

export interface YoutubePlaylistBody {
  /** Full YouTube video or playlist URL (e.g. youtube.com/watch?v=xxx or youtube.com/playlist?list=PLxxx) */
  url: string;
  /** Optional custom title (defaults to the title from YouTube) */
  title?: string;
  /** When true (default), fetches a transcript for each video in a playlist. Set to false to skip transcript fetching and speed up import. */
  fetchTranscripts?: boolean;
}

export type NodeType = (typeof NodeType)[keyof typeof NodeType];

export const NodeType = {
  domain: "domain",
  concept: "concept",
  insight: "insight",
  action: "action",
  goal: "goal",
} as const;

/**
 * Difficulty/progression level — only set on concept and goal nodes; null for domain, insight, and action nodes (or for pre-existing nodes that have not been re-processed)
 */
export type NodeLevel = (typeof NodeLevel)[keyof typeof NodeLevel] | null;

export const NodeLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export type NodeSourceType =
  (typeof NodeSourceType)[keyof typeof NodeSourceType];

export const NodeSourceType = {
  text: "text",
  url: "url",
  note: "note",
  file: "file",
  youtube: "youtube",
  instagram: "instagram",
  pinterest: "pinterest",
  image: "image",
  video: "video",
  audio: "audio",
} as const;

export interface NodeSource {
  inputId: string;
  title: string;
  type: NodeSourceType;
}

export interface Node {
  id: string;
  label: string;
  type: NodeType;
  /** Difficulty/progression level — only set on concept and goal nodes; null for domain, insight, and action nodes (or for pre-existing nodes that have not been re-processed) */
  level?: NodeLevel;
  description?: string | null;
  inputId?: string | null;
  /** Number of distinct source inputs that contributed to this node (via merging) */
  sourceCount?: number;
  /** Full source list — only populated by GET /nodes/:id, null elsewhere */
  sources?: NodeSource[] | null;
  positionX: number;
  positionY: number;
  createdAt: string;
  updatedAt: string;
}

export type ActionStatus = (typeof ActionStatus)[keyof typeof ActionStatus];

export const ActionStatus = {
  pending: "pending",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface Action {
  id: string;
  title: string;
  description?: string | null;
  status: ActionStatus;
  nodeId?: string | null;
  /** The source input that generated this action */
  inputId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Edge {
  id: string;
  sourceId: string;
  targetId: string;
  label?: string | null;
  createdAt: string;
}

export interface ProcessResult {
  input: Input;
  nodesCreated: Node[];
  actionsCreated: Action[];
  edgesCreated: Edge[];
}

export type CreateNodeBodyType =
  (typeof CreateNodeBodyType)[keyof typeof CreateNodeBodyType];

export const CreateNodeBodyType = {
  domain: "domain",
  concept: "concept",
  insight: "insight",
  action: "action",
  goal: "goal",
} as const;

export interface CreateNodeBody {
  label: string;
  type: CreateNodeBodyType;
  description?: string;
  inputId?: string;
  positionX?: number;
  positionY?: number;
}

export type UpdateNodeBodyType =
  (typeof UpdateNodeBodyType)[keyof typeof UpdateNodeBodyType];

export const UpdateNodeBodyType = {
  domain: "domain",
  concept: "concept",
  insight: "insight",
  action: "action",
  goal: "goal",
} as const;

export interface UpdateNodeBody {
  label?: string;
  type?: UpdateNodeBodyType;
  description?: string;
  positionX?: number;
  positionY?: number;
}

export interface CreateEdgeBody {
  sourceId: string;
  targetId: string;
  label?: string;
}

export type UpdateActionBodyStatus =
  (typeof UpdateActionBodyStatus)[keyof typeof UpdateActionBodyStatus];

export const UpdateActionBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface UpdateActionBody {
  title?: string;
  description?: string;
  status?: UpdateActionBodyStatus;
}

export type GraphSummaryNodesByType = {
  domain: number;
  concept: number;
  insight: number;
  action: number;
  goal: number;
};

export type GraphSummaryActionsByStatus = {
  pending: number;
  in_progress: number;
  done: number;
};

export interface GraphSummary {
  totalNodes: number;
  totalEdges: number;
  totalInputs: number;
  totalActions: number;
  nodesByType: GraphSummaryNodesByType;
  actionsByStatus: GraphSummaryActionsByStatus;
}

export interface GraphData {
  nodes: Node[];
  edges: Edge[];
}

export interface GraphDomains {
  /** Alphabetically sorted list of distinct domain labels in the graph */
  domains: string[];
}

export interface FileUploadError {
  /** Original filename that failed */
  file: string;
  /** Human-readable reason for the failure */
  error: string;
}

export interface UploadResult {
  /** Inputs successfully created from uploaded files */
  created: Input[];
  /** Files that could not be processed */
  errors: FileUploadError[];
}

export interface ApiError {
  error: string;
}

export type UploadFilesBody = {
  files: Blob[];
};