# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Headmaster API - Personal AI Knowledge System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nodes:
    get:
      operationId: listNodes
      tags: [nodes]
      summary: List all nodes
      responses:
        "200":
          description: List of nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
    post:
      operationId: createNode
      tags: [nodes]
      summary: Create a new node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeBody"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
  /nodes/{id}:
    get:
      operationId: getNode
      tags: [nodes]
      summary: Get a single node
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Node found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
    put:
      operationId: updateNode
      tags: [nodes]
      summary: Update a node
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeBody"
      responses:
        "200":
          description: Node updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
    delete:
      operationId: deleteNode
      tags: [nodes]
      summary: Delete a node
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /edges:
    get:
      operationId: listEdges
      tags: [edges]
      summary: List all edges
      responses:
        "200":
          description: List of edges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Edge"
    post:
      operationId: createEdge
      tags: [edges]
      summary: Create an edge between nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEdgeBody"
      responses:
        "201":
          description: Edge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edge"
  /edges/{id}:
    delete:
      operationId: deleteEdge
      tags: [edges]
      summary: Delete an edge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /actions/{id}:
    put:
      operationId: updateAction
      tags: [actions]
      summary: Update an action item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActionBody"
      responses:
        "200":
          description: Action updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
    delete:
      operationId: deleteAction
      tags: [actions]
      summary: Delete an action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    NodeSource:
      type: object
      properties:
        inputId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [text, url, note, file, youtube, instagram, pinterest, image, video, audio]
      required:
        - inputId
        - title
        - type
    Node:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [domain, concept, insight, action, goal]
        level:
          type: string
          enum: [beginner, intermediate, advanced]
          nullable: true
          description: Difficulty/progression level — only set on concept and goal nodes; null for domain, insight, and action nodes (or for pre-existing nodes that have not been re-processed)
        description:
          type: string
          nullable: true
        inputId:
          type: string
          nullable: true
        sourceCount:
          type: integer
          description: Number of distinct source inputs that contributed to this node (via merging)
        sources:
          type: array
          nullable: true
          description: Source inputs that contributed to this node — included in both GET /graph/data and GET /nodes/:id
          items:
            $ref: "#/components/schemas/NodeSource"
        positionX:
          type: number
        positionY:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - label
        - type
        - positionX
        - positionY
        - createdAt
        - updatedAt
    CreateNodeBody:
      type: object
      properties:
        label:
          type: string
        type:
          type: string
          enum: [domain, concept, insight, action, goal]
        description:
          type: string
        inputId:
          type: string
        positionX:
          type: number
        positionY:
          type: number
      required:
        - label
        - type
    UpdateNodeBody:
      type: object
      properties:
        label:
          type: string
        type:
          type: string
          enum: [domain, concept, insight, action, goal]
        description:
          type: string
        positionX:
          type: number
        positionY:
          type: number
    Edge:
      type: object
      properties:
        id:
          type: string
        sourceId:
          type: string
        targetId:
          type: string
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sourceId
        - targetId
        - createdAt
    CreateEdgeBody:
      type: object
      properties:
        sourceId:
          type: string
        targetId:
          type: string
        label:
          type: string
      required:
        - sourceId
        - targetId
    Action:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, done]
        nodeId:
          type: string
          nullable: true
        inputId:
          type: string
          nullable: true
          description: The source input that generated this action
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - createdAt
        - updatedAt
    UpdateActionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, done]
...