# lib/api-zod/src/generated/types/edge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Headmaster API - Personal AI Knowledge System
 * OpenAPI spec version: 0.1.0
 */

export interface Edge {
  id: string;
  sourceId: string;
  targetId: string;
  label?: string | null;
  createdAt: Date;
}