# lib/api-zod/src/generated/types/node.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Headmaster API - Personal AI Knowledge System
 * OpenAPI spec version: 0.1.0
 */
import type { NodeLevel } from "./nodeLevel";
import type { NodeSource } from "./nodeSource";
import type { NodeType } from "./nodeType";

export interface Node {
  id: string;
  label: string;
  type: NodeType;
  /** Difficulty/progression level — only set on concept and goal nodes; null for domain, insight, and action nodes (or for pre-existing nodes that have not been re-processed) */
  level?: NodeLevel;
  description?: string | null;
  inputId?: string | null;
  /** Number of distinct source inputs that contributed to this node (via merging) */
  sourceCount?: number;
  /** Full source list — only populated by GET /nodes/:id, null elsewhere */
  sources?: NodeSource[] | null;
  positionX: number;
  positionY: number;
  createdAt: Date;
  updatedAt: Date;
}